+++ front-end/src/Views/LoginPage/style.css
@@ -17,16 +17,6 @@
   align-items: flex-start;
 }
 
-.loginPage__bgImage {
-  position: absolute;
-  max-width: 100%;
-  /* max-height: 100vh; */
-  opacity: 0.3;
-  /* top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%); */
-}
-
 .loginPage__header {
   text-align: center;
 }
